Create DynamoDB table, movie attribute list
* Co-authored by: Alec Dutcher @aidutcher

Create a DynamoDB table to store the movie_content we have. Our thought
process here is to use imdbID as a unique primary key, as that's already
unique enough and we don't need more sorting.

We also changed the working number of titles to four, to avoid hitting
Developer API limits with OMDB.

Next session we'll work on trying to get content into Dynamo, and are
thinking of expeirmenting with data analysis via recommendation engine.
We briefly looked at Amazon Personalize for this.